slapnotes/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
django-dynamic-preferences==1.7
django-filter==2.1.0
django-filters==0.2.1
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
persisting-theory==0.2.1
pkg-resources==0.0.0
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24.1
